fix classification for binary format (ref: 9834a0ce49bd098535c276358015b2c64e20696a)

+++ src/workers/BinaryDecoderWorker.js
@@ -116,7 +116,7 @@ self.onmessage = function (event) {
 				for (let j = 0; j < numPoints; j++) {
 					let classification = cv.getUint8(inOffset + j * pointAttributes.byteSize);
 					let firstByte = j * iStride + outOffset;
-					iView.setUint8(firstByte + 0, classification, true);
+					iView.setFloat32(firstByte + 0, classification, true);
 				}
 			} else if (pointAttribute.name === PointAttribute.NORMAL_SPHEREMAPPED.name) {
 				for (let j = 0; j < numPoints; j++) {
